real cpu implementation
Old is obsolete, real code is in src
